вынес презентер точки маршрута из главного презентера
