+++ src/model/points-model.js
@@ -1,6 +1,6 @@
 import { getMockPoint } from '../mock/waypoint.js';
 
-const POINT_COUNT = 0;
+const POINT_COUNT = 4;
 
 export default class PointsModel {
   #points = Array.from({ length: POINT_COUNT }, () => getMockPoint());
