+++ src/presenter/trip-presenter.js
@@ -1,11 +1,10 @@
 import TripInfoView from '../view/trip-info-view';
 import FilterView from '../view/filter-view';
 import SortView from '../view/sort-view';
-import PointView from '../view/point-view';
 import PointsListView from '../view/points-list-view';
-import PointEditView from '../view/point-edit-view';
-import {render, RenderPosition, replace} from '../framework/render.js';
+import {render, RenderPosition} from '../framework/render.js';
 import EmptyListView from '../view/empty-list-view';
+import PointPresenter from './point-presenter';
 
 
 export default class TripPresenter {
@@ -19,58 +18,35 @@ export default class TripPresenter {
   }
 
   init() {
-    this.#tripPoints = [...this.#pointsModel.points];
-
     this.#renderTrip();
   }
 
   #renderTrip = () => {
-    render(new FilterView(), this.#containers.filterContainer);
-    if (this.#tripPoints.length > 0) {
-      render(new TripInfoView(), this.#containers.tripInfoContainer, RenderPosition.AFTERBEGIN);
-      render(new SortView(), this.#containers.eventContainer);
-      render(this.#listPoints, this.#containers.eventContainer);
+    this.#tripPoints = [...this.#pointsModel.points];
 
-      this.#tripPoints.forEach((point) => this.#renderPoint(point));
+    render(new FilterView(), this.#containers.filterContainer);
+
+    if (this.#tripPoints.length === 0) {
+      this.#renderEmptyList();
+      return;
     }
-    else{
-      render(new EmptyListView(), this.#containers.eventContainer);
-    }
+
+    render(new TripInfoView(), this.#containers.tripInfoContainer, RenderPosition.AFTERBEGIN);
+    render(new SortView(), this.#containers.eventContainer);
+    render(this.#listPoints, this.#containers.eventContainer);
+
+    this.#tripPoints.forEach((point) => this.#renderPoint(point));
   };
 
   #renderPoint = (point) => {
-    const escKeyHandler = (evt) => {
-      if (evt.key === 'Escape') {
-        evt.preventDefault();
-        replaceEditToPointView();
-        document.removeEventListener('keydown', escKeyHandler);
-      }
-    };
-
-    const pointComponent = new PointView({point: point, onEditClick: () => {
-      replacePointToEditView();
-      document.addEventListener('keydown', escKeyHandler);
-    }});
-
-    const editPointComponent = new PointEditView({point: point,
-      onSubmitClick: () => {
-        replaceEditToPointView();
-        document.addEventListener('keydown', escKeyHandler);
-      },
-      onRollUpClick: () => {
-        replaceEditToPointView();
-        document.addEventListener('keydown', escKeyHandler);
-      }
+    const pointPresenter = new PointPresenter({
+      point: point,
+      container: this.#listPoints,
     });
+    pointPresenter.init();
+  };
 
-    function replacePointToEditView() {
-      replace(editPointComponent, pointComponent);
-    }
-
-    function replaceEditToPointView() {
-      replace(pointComponent, editPointComponent);
-    }
-
-    render(pointComponent, this.#listPoints.element);
+  #renderEmptyList = () => {
+    render(new EmptyListView(), this.#containers.eventContainer);
   };
 }
